moved rest methods to service
