+++ src/main/java/org/ohdsi/webapi/ircalc/IRExecutionInfoRepository.java
@@ -1,7 +1,6 @@
 package org.ohdsi.webapi.ircalc;
 
 import org.ohdsi.webapi.GenerationStatus;
-import org.ohdsi.webapi.source.Source;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
@@ -9,5 +8,4 @@ import java.util.List;
 public interface IRExecutionInfoRepository extends CrudRepository<ExecutionInfo, ExecutionInfoId> {
     List<ExecutionInfo> findByStatus(GenerationStatus status);
     List<ExecutionInfo> findByStatusIn(List<GenerationStatus> statuses);
-    List<ExecutionInfo> findBySource(Source source);
 }
