+++ src/main/java/org/ohdsi/webapi/service/SourceService.java
@@ -26,8 +26,6 @@ import org.glassfish.jersey.media.multipart.FormDataParam;
 import org.jasypt.encryption.pbe.PBEStringEncryptor;
 import org.jasypt.properties.PropertyValueEncryptionUtils;
 import org.ohdsi.sql.SqlTranslate;
-import org.ohdsi.webapi.ircalc.ExecutionInfo;
-import org.ohdsi.webapi.ircalc.IRExecutionInfoRepository;
 import org.ohdsi.webapi.shiro.management.Security;
 import org.ohdsi.webapi.source.*;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -61,8 +59,6 @@ public class SourceService extends AbstractDaoService {
   private Environment env;
   @Autowired
   private ApplicationEventPublisher publisher;
-  @Autowired
-  private IRExecutionInfoRepository irExecutionInfoRepository;
   @Value("${datasource.ohdsi.schema}")
   private String schema;
